move writing priority from server to spring-boot
