+++ generators/spring-boot/cleanup-oauth2.ts
@@ -16,5 +16,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default } from './generator.js';
-export { default as command } from './command.js';
+import { asWritingTask } from '../base-application/support/task-type-inference.js';
+
+/**
+ * Removes server files that where generated in previous JHipster versions and therefore
+ * need to be removed.
+ */
+export default asWritingTask(function cleanupOldServerFilesTask(this, { application }) {
+  if (this.isJhipsterVersionLessThan('6.0.0')) {
+    this.removeFile(`${application.javaPackageSrcDir}config/OAuth2Configuration.java`);
+    this.removeFile(`${application.javaPackageSrcDir}security/OAuth2AuthenticationSuccessHandler.java`);
+  }
+  if (this.isJhipsterVersionLessThan('7.6.1')) {
+    if (!application.databaseTypeNo) {
+      this.removeFile(`${application.javaPackageSrcDir}web/rest/UserResource.java`);
+    }
+  }
+});
